Optimizing user deletion

+++ eupheus/app/Http/Controllers/LA/UsersController.php
@@ -196,8 +196,14 @@ class UsersController extends Controller
 	{
 		if(Module::hasAccess("Users", "delete")) {
 			User::find($id)->forceDelete();
-			Student::find($id)->forceDelete();
-			Employee::find($id)->forceDelete();
+
+			if(Student::find($id)->exists()) {
+				Student::find($id)->forceDelete();
+			}
+
+			if(Employee::find($id)->exists()) {
+				Employee::find($id)->forceDelete();
+			}
 			
 			// Redirecting to index() method
 			return redirect()->route(config('laraadmin.adminRoute') . '.users.index');
